station health, player health and boost bars added

+++ src/gameplay/station.cpp
@@ -9,7 +9,7 @@ namespace se {
 		GLuint t_program, 
 		const bounding_sphere& t_sphere,
 		const obb& t_box,
-		float t_health
+		int t_health
 	) : transformable(t_position, {1,0,0,0}),
 		renderable(t_position, {1,0,0,0}, t_mesh, t_textures, t_program),
 		collidable(t_sphere, t_box),
